Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-dotenv
textblob
openai==0.28
pymongo
motor
pydantic[email]

